+++ e2e/custom-levels/types.d.ts
@@ -1,0 +1,5 @@
+declare module 'electron-log' {
+  interface LogFunctions {
+    notice(...params: any[]): void;
+  }
+}